reverse-etl/webapp/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
azure-cosmos
azure-identity
openai>=1.0.0
